fix: refactor reset disk and finch config test (#599)
Issue #, if available:

*Description of changes:*
- fix path for deleting persistent disk for windows for cleanup in e2e
tests
- fix path for finch config e2e vm test
*Testing done:*

Yes, locally

- [X] I've reviewed the guidance in CONTRIBUTING.md


#### License Acceptance

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

---------
