+++ e2e/vm/vm_darwin_test.go
@@ -7,12 +7,18 @@
 package vm
 
 import (
+	"errors"
+	"io/fs"
+	"os"
+	"os/exec"
+	"path/filepath"
 	"testing"
 
 	"github.com/onsi/ginkgo/v2"
 	"github.com/onsi/gomega"
 
 	"github.com/runfinch/common-tests/command"
+	"github.com/runfinch/common-tests/option"
 
 	"github.com/runfinch/finch/e2e"
 )
@@ -50,4 +56,25 @@ func TestVM(t *testing.T) {
 
 	gomega.RegisterFailHandler(ginkgo.Fail)
 	ginkgo.RunSpecs(t, description)
+}
+
+var resetDisks = func(o *option.Option, installed bool) {
+	var dataDiskDir string
+	limaDisksPath := "lima/data/_disks/"
+	if installed {
+		path, err := exec.LookPath(e2e.InstalledTestSubject)
+		gomega.Expect(err).ShouldNot(gomega.HaveOccurred())
+		realFinchPath, err := filepath.EvalSymlinks(path)
+		gomega.Expect(err).ShouldNot(gomega.HaveOccurred())
+		dataDiskDir = filepath.Join(realFinchPath, "..", "..", limaDisksPath)
+	} else {
+		dataDiskDir = filepath.Join("..", "..", "_output", limaDisksPath)
+	}
+	realDiskPath, err := os.Readlink(filepath.Join(dataDiskDir, "finch", "datadisk"))
+	if err == nil {
+		gomega.Expect(os.Remove(realDiskPath)).ShouldNot(gomega.HaveOccurred())
+	} else if !errors.Is(err, fs.ErrNotExist) {
+		gomega.Expect(err).ShouldNot(gomega.HaveOccurred())
+	}
+	gomega.Expect(os.RemoveAll(dataDiskDir)).ShouldNot(gomega.HaveOccurred())
 }
