+++ e2e/vm/vm_test.go
@@ -4,9 +4,6 @@
 package vm
 
 import (
-	"errors"
-	"io/fs"
-	"os"
 	"os/exec"
 	"path/filepath"
 
@@ -32,7 +29,7 @@ var resetVM = func(o *option.Option, installed bool) string {
 		gomega.Expect(err).ShouldNot(gomega.HaveOccurred())
 		realFinchPath, err := filepath.EvalSymlinks(path)
 		gomega.Expect(err).ShouldNot(gomega.HaveOccurred())
-		limaConfigFilePath = filepath.Join(realFinchPath, "../../lima/data/_config/override.yaml")
+		limaConfigFilePath = filepath.Join(realFinchPath, "..", "..", "lima", "data", "_config", "override.yaml")
 	}
 	origLimaCfg := readFile(limaConfigFilePath)
 
@@ -48,25 +45,4 @@ var resetVM = func(o *option.Option, installed bool) string {
 	})
 
 	return limaConfigFilePath
-}
-
-var resetDisks = func(o *option.Option, installed bool) {
-	var dataDiskDir string
-	limaDisksPath := "lima/data/_disks/"
-	if installed {
-		path, err := exec.LookPath(e2e.InstalledTestSubject)
-		gomega.Expect(err).ShouldNot(gomega.HaveOccurred())
-		realFinchPath, err := filepath.EvalSymlinks(path)
-		gomega.Expect(err).ShouldNot(gomega.HaveOccurred())
-		dataDiskDir = filepath.Join(realFinchPath, "..", "..", limaDisksPath)
-	} else {
-		dataDiskDir = filepath.Join("..", "..", "_output/", limaDisksPath)
-	}
-	realDiskPath, err := os.Readlink(filepath.Join(dataDiskDir, "finch", "datadisk"))
-	if err == nil {
-		gomega.Expect(os.Remove(realDiskPath)).ShouldNot(gomega.HaveOccurred())
-	} else if !errors.Is(err, fs.ErrNotExist) {
-		gomega.Expect(err).ShouldNot(gomega.HaveOccurred())
-	}
-	gomega.Expect(os.RemoveAll(dataDiskDir)).ShouldNot(gomega.HaveOccurred())
 }
